Added amount for products, dates to rating and comment, admin panel

+++ apps/shop/models/order.py
@@ -2,6 +2,7 @@ from django.db import models
 from apps.users.models.user import User
 from apps.shop.models.pick_up_point import PickUpModel
 from django.core.exceptions import ValidationError
+from django.forms import ModelForm
 
 class OrderModel(models.Model):
     AM_9_10 = "9.00-10.00"
@@ -34,8 +35,5 @@ class OrderModel(models.Model):
     date_of_pick_up = models.DateField(null = False)
     time_of_pick_up = models.CharField (choices=TIME_IN_PICK_UP,max_length = 100)
     comment = models.TextField(null = True, default = 'Order without Comment')
-    total_money = models.FloatField(null=False)
-    
-    # def validate_data (self):
-    #     if len(OrderModel.objects.filter(pick_up_point = self.pick_up_point).filter(date_of_pick_up=self.date_of_pick_up).filter(time_of_pick_up=self.time_of_pick_up)) >= 5:
-    #         return ValidationError
+    total_money = models.DecimalField(max_digits=10,decimal_places=2,null=False)
+
